requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Limiter==4.0.0
python-socketio==5.11.1
requests==2.31.0
PyYAML==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
mypy==1.7.1
flasgger==0.9.7.1
